@@ -4,32 +4,35 @@ publish_to: 'none'
 version: 0.0.1
 
 environment:
-  sdk: ">=3.5.4 <4.0.0"
-  flutter: "3.24.5"
+  sdk: ">=3.6.2 <4.0.0"
+  flutter: "3.29.1"
+
+resolution: workspace
 
 dependencies:
-  device_info_plus: 11.2.0
+  device_info_plus: 11.3.3
   flutter:
     sdk: flutter
-  flutter_secure_storage: 9.2.2
+  flutter_secure_storage: 9.2.4
   foundation:
     path: ../foundation
   fpdart: 1.1.1
-  freezed_annotation: 2.4.4
+  freezed_annotation: 3.0.0
   get_it: 8.0.3
   json_annotation: 4.9.0
-  json_serializable: 6.9.0
+  json_serializable: 6.9.4
   project_linter:
     path: ../project_linter
   rxdart: 0.28.0
-  shared_preferences: 2.3.4
+  shared_preferences: 2.5.2
 
 dev_dependencies:
-  analyzer: 6.7.0
-  build_runner: 2.4.13
-  freezed: 2.5.7
+  analyzer: 7.3.0
+  build_runner: 2.4.15
+  flutter_test:
+    sdk: flutter
+  freezed: 3.0.3
   mocktail: 1.0.4
   platform: 3.1.6
-  test: 1.25.14
 
 flutter:
