@@ -4,14 +4,16 @@ publish_to: 'none'
 version: 0.0.1
 
 environment:
-  sdk: ">=3.5.4 <4.0.0"
-  flutter: "3.24.5"
+  sdk: ">=3.6.2 <4.0.0"
+  flutter: "3.29.1"
+
+resolution: workspace
 
 dependencies:
   flutter:
     sdk: flutter
   fpdart: 1.1.1
-  logging: 1.1.0
+  logging: 1.3.0
   project_linter:
     path: ../project_linter
 
