@@ -4,12 +4,14 @@ publish_to: 'none'
 version: 0.0.1
 
 environment:
-  sdk: ">=3.5.4 <4.0.0"
-  flutter: "3.24.5"
+  sdk: ">=3.6.2 <4.0.0"
+  flutter: "3.29.1"
+
+resolution: workspace
 
 dependencies:
-  collection: 1.18.0
-  dio: 5.7.0
+  collection: 1.19.1
+  dio: 5.8.0+1
   domain:
     path: ../domain
   flutter:
@@ -18,20 +20,21 @@ dependencies:
   foundation:
     path: ../foundation
   fpdart: 1.1.1
-  freezed_annotation: 2.4.4
+  freezed_annotation: 3.0.0
   get_it: 8.0.3
   json_annotation: 4.9.0
-  json_serializable: 6.9.0
+  json_serializable: 6.9.4
   project_linter:
     path: ../project_linter
-  retrofit: 4.4.1
+  retrofit: 4.4.2
 
 dev_dependencies:
-  build_runner: 2.4.13
-  freezed: 2.5.7
+  build_runner: 2.4.15
+  flutter_test:
+    sdk: flutter
+  freezed: 3.0.3
   mocktail: 1.0.4
   platform: 3.1.6
-  retrofit_generator: 9.1.5
-  test: 1.25.14
+  retrofit_generator: 9.1.9
 
 flutter:
