@@ -4,7 +4,7 @@ part '{{domainModelName.snakeCase()}}.freezed.dart';
 part '{{domainModelName.snakeCase()}}.g.dart';
 
 @freezed
-class {{domainModelName.pascalCase()}} with _${{domainModelName.pascalCase()}} {
+abstract class {{domainModelName.pascalCase()}} with _${{domainModelName.pascalCase()}} {
   const factory {{domainModelName.pascalCase()}}({
     // put your domain model fields here
   }) = _{{domainModelName.pascalCase()}};
