@@ -4,7 +4,7 @@ part '{{usecaseName.snakeCase()}}_request_remote_model.freezed.dart';
 part '{{usecaseName.snakeCase()}}_request_remote_model.g.dart';
 
 @freezed
-class {{usecaseName.pascalCase()}}RequestRemoteModel with _${{usecaseName.pascalCase()}}RequestRemoteModel {
+abstract class {{usecaseName.pascalCase()}}RequestRemoteModel with _${{usecaseName.pascalCase()}}RequestRemoteModel {
   const factory {{usecaseName.pascalCase()}}RequestRemoteModel({
     // put request remote model fields here
   }) = _{{usecaseName.pascalCase()}}RequestRemoteModel;
