@@ -1,6 +1,6 @@
 part of '{{name.snakeCase()}}_bloc.dart';
 
 @freezed
-class {{name.pascalCase()}}Event with _${{name.pascalCase()}}Event {
+abstract class {{name.pascalCase()}}Event with _${{name.pascalCase()}}Event {
   const factory {{name.pascalCase()}}Event.onInitiated() = _OnInitiated;
 }
