@@ -1,7 +1,7 @@
 part of '{{name.snakeCase()}}_bloc.dart';
 
 @freezed
-class {{name.pascalCase()}}State with _${{name.pascalCase()}}State {
+abstract class {{name.pascalCase()}}State with _${{name.pascalCase()}}State {
   const factory {{name.pascalCase()}}State({
     required StateType type,
     required {{name.pascalCase()}}Argument argument,
