@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'error_detail.freezed.dart';
 
 @freezed
-class ErrorDetail with _$ErrorDetail {
+abstract class ErrorDetail with _$ErrorDetail {
   const factory ErrorDetail.backend({
     String? errorCode,
     String? message,
