@@ -5,7 +5,7 @@ part 'user.freezed.dart';
 part 'user.g.dart';
 
 @freezed
-class User with _$User implements Storable {
+abstract class User with _$User implements Storable {
   static const loggedInUserStoreKey = 'loggedInUserStore';
 
   const factory User({
