@@ -1,7 +1,7 @@
 import 'package:domain/store/adapter/secure_storage_adapter.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:test/test.dart';
 
 class MockFlutterSecureStorage extends Mock implements FlutterSecureStorage {}
 
