@@ -1,7 +1,7 @@
 part of 'auth_bloc.dart';
 
 @freezed
-class AuthEvent with _$AuthEvent {
+abstract class AuthEvent with _$AuthEvent {
   const factory AuthEvent.onInitiated() = _OnInitiated;
   const factory AuthEvent.onUserUpdated(User? user) = _OnUserUpdated;
   const factory AuthEvent.onAuthCheckRequested() = _OnAuthCheckRequested;
