@@ -14,32 +14,30 @@ class SnackBarContainer extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocListener<SnackBarBloc, SnackBarState>(
       listener: (context, state) {
-        state.mapOrNull(
-          showMessage: (it) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                backgroundColor: it.snackBarMessage.background ?? context.palette.snackBarColor,
-                content: Row(
-                  children: [
-                    if (it.snackBarMessage.icon != null) ...[
-                      it.snackBarMessage.icon!,
-                      const SizedBox(width: Insets.xSmall),
-                    ],
-                    Expanded(child: Text(it.snackBarMessage.message)),
+        if (state is SnackBarShowMessage) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              backgroundColor: state.snackBarMessage.background ?? context.palette.snackBarColor,
+              content: Row(
+                children: [
+                  if (state.snackBarMessage.icon != null) ...[
+                    state.snackBarMessage.icon!,
+                    const SizedBox(width: Insets.xSmall),
                   ],
-                ),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(10),
-                ),
-                padding: const EdgeInsets.symmetric(horizontal: Insets.large, vertical: Insets.xLarge),
-                margin: const EdgeInsets.symmetric(horizontal: Insets.xxxLarge, vertical: Insets.large),
-                elevation: 0,
-                behavior: SnackBarBehavior.floating,
-                dismissDirection: DismissDirection.startToEnd,
+                  Expanded(child: Text(state.snackBarMessage.message)),
+                ],
               ),
-            );
-          },
-        );
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10),
+              ),
+              padding: const EdgeInsets.symmetric(horizontal: Insets.large, vertical: Insets.xLarge),
+              margin: const EdgeInsets.symmetric(horizontal: Insets.xxxLarge, vertical: Insets.large),
+              elevation: 0,
+              behavior: SnackBarBehavior.floating,
+              dismissDirection: DismissDirection.startToEnd,
+            ),
+          );
+        }
       },
       child: child,
     );
