@@ -1,6 +1,6 @@
 part of 'dashboard_bloc.dart';
 
 @freezed
-class DashboardEvent with _$DashboardEvent {
+abstract class DashboardEvent with _$DashboardEvent {
   const factory DashboardEvent.onInitiated() = _OnInitiated;
 }
