@@ -1,7 +1,7 @@
 part of 'dashboard_bloc.dart';
 
 @freezed
-class DashboardState with _$DashboardState {
+abstract class DashboardState with _$DashboardState {
   const factory DashboardState({
     required StateType type,
     required DashboardArgument argument,
