@@ -1,6 +1,6 @@
 part of 'dashboard_home_bloc.dart';
 
 @freezed
-class DashboardHomeEvent with _$DashboardHomeEvent {
+abstract class DashboardHomeEvent with _$DashboardHomeEvent {
   const factory DashboardHomeEvent.onInitiated() = _OnInitiated;
 }
