@@ -1,6 +1,6 @@
 part of 'dashboard_profile_bloc.dart';
 
 @freezed
-class DashboardProfileEvent with _$DashboardProfileEvent {
+abstract class DashboardProfileEvent with _$DashboardProfileEvent {
   const factory DashboardProfileEvent.onInitiated() = _OnInitiated;
 }
