@@ -1,7 +1,7 @@
 part of 'dashboard_profile_bloc.dart';
 
 @freezed
-class DashboardProfileState with _$DashboardProfileState {
+abstract class DashboardProfileState with _$DashboardProfileState {
   const factory DashboardProfileState({
     required StateType type,
     required DashboardProfileArgument argument,
