@@ -1,7 +1,7 @@
 part of 'welcome_bloc.dart';
 
 @freezed
-class WelcomeEvent with _$WelcomeEvent {
+abstract class WelcomeEvent with _$WelcomeEvent {
   const factory WelcomeEvent.onInitiated() = _OnInitiated;
   const factory WelcomeEvent.onLoggedIn() = _OnLoggedIn;
 }
