@@ -1,5 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
 import 'package:remote/models/user/user_remote_model.dart';
-import 'package:test/test.dart';
 
 void main() {
   group(
@@ -19,7 +19,7 @@ void main() {
 
       test(
         'should create user remote model from json',
-        () async {
+        () {
           final result = UserRemoteModel.fromJson(jsonMap);
 
           expect(result, user);
@@ -28,7 +28,7 @@ void main() {
 
       test(
         'should create json from user remote model',
-        () async {
+        () {
           final result = user.toJson();
 
           expect(result, jsonMap);
